+++ Tests/ReactiveSwiftTests/ActionSpec.swift
@@ -411,6 +411,51 @@ class ActionSpec: QuickSpec {
 
 				expect(values) == [3, 4, -11]
 			}
+			
+			it("is disabled if the validating property does not hold a valid value") {
+				enum TestValidationError: Error { case generic }
+				typealias PropertyType = ValidatingProperty<Int, TestValidationError>
+				let decisions: [PropertyType.Decision] = [.valid, .invalid(.generic), .coerced(10, nil)]
+				let input = PropertyType(0, { decisions[$0] })
+				let action = Action(validated: input, execute: echo)
+				expect(action.isEnabled.value) == true
+				expect(action.apply().single()?.value) == 0
+				input.value = 1
+				expect(action.isEnabled.value) == false
+				expect(action.apply().single()?.error).toNot(beNil())
+				input.value = 2
+				expect(action.isEnabled.value) == true
+				expect(action.apply().single()?.value) == 10
+			}
+			
+			it("allows a different input type while using a validating property as its state") {
+				enum TestValidationError: Error { case generic }
+				let state = ValidatingProperty<Int?, TestValidationError>(nil, { $0 != nil ? .valid : .invalid(.generic) })
+				
+				let add = Action<String, Int?, NoError>(validated: state) { state, input -> SignalProducer<Int?, NoError> in
+					guard let input = Int(input), let state = state else { return SignalProducer(value: nil) }
+					return SignalProducer(value: state + input)
+				}
+				
+				var values: [Int] = []
+				add.values.observeValues { output in
+					if let output = output {
+						values.append(output)
+					}
+				}
+				
+				expect(add.isEnabled.value) == false
+				state.value = 1
+				expect(add.isEnabled.value) == true
+				
+				add.apply("2").start()
+				add.apply("3").start()
+				
+				state.value = -1
+				add.apply("-10").start()
+				
+				expect(values) == [3, 4, -11]
+			}
 		}
 	}
 }