+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # master
+*Please add new entries at the top.*
 
+1. New convenience initializer for `Action` that takes a `ValidatingProperty` as its state (#637, kudos to @Marcocanc)
 
-# 4.0.0*Please add new entries at the top.*
+# 4.0.0
 
 1. When unfair locks from libplatform are unavailable, ReactiveSwift now fallbacks to error checking Pthread mutexes instead of the default. Mitigations regarding issues with `pthread_mutex_trylock` have also been applied. (#654, kudos to @andersio)
 1. Fix some documentation errors about Carthage usage (#655)